# webapp/src/main/webapp/docs/swagger.yaml
openapi: 3.0.1
info:
  title: HealthSync API
  version: 1.0.0
...
paths:
  /appointments:
    get:
      operationId: getAppointments
      description: Gets a page of the user's appointments
      tags:
        - appointments
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            minimum: 0
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            type: string
            enum: [CONFIRMED, CANCELLED, COMPLETED]
          example: COMPLETED
        - name: page
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: Returns appointment Page
          content:
            application/vnd.appointment-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Appointment'
        '204':
          description: No content for specified page
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '403':
          description: Only the user can see its appointments
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'

    post:
      operationId: createAppointment
      description: Creates an appointment
      tags:
        - appointments
      requestBody:
        content:
          application/vnd.appointment.v1+json:
            schema:
              $ref : '#/components/schemas/AppointmentForm'
      responses:
        '201':
          description: Appointment created successfully
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '403':
          description: Only patients can request appointments
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Doctor not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '409':
          description: Doctor not available at specified date and time
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
  /appointments/{appointmentId}:
    get:
      operationId: getAppointment
      description: Returns the appointment
      tags:
        - appointments
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returns the appointment
          content:
            application/vnd.appointment.v1+json:
              schema:
                allOf:
                  - $ref : '#/components/schemas/Appointment'
        '403':
          description: Only the users involved in the appointment can see it
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Appointment not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'

    patch:
      operationId: cancelAppointment
      description: Cancels an appointment
      tags:
        - appointments
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/vnd.appointment-cancel.v1+json:
            schema:
              $ref: '#/components/schemas/CancelAppointmentForm'
      responses:
        '200':
          description: Appointment cancelled
          content:
            application/vnd.appointment.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Appointment'
              example:
                {
                  "cancelDescription": "Tengo examen ese dia, mas adelante agendo otro turno",
                  "date": "2023-05-20",
                  "description": "Me duele el pie",
                  "doctor": "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20",
                  "id": 12,
                  "patient": "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/patients/21",
                  "self": "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/appointments/12",
                  "status": "CANCELLED",
                  "timeBlock": "10:30"
                }
        '403':
          description: Only the users involved in the appointment can cancel it
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Appointment not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '409':
          description: Appointment is in an inmutable state
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
...
  /appointments/{appointmentId}/indications:
    get:
      operationId: listIndications
      description: Returns a page of the appointment's indications
      tags:
        - indications
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: Returned indication page
          content:
            application/vnd.indication-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref : '#/components/schemas/Indication'
        '204':
          description: No content for specified page
        '403':
          description: Only the patient or the doctor can see the appointment's indications
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Appointment not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
    post:
      operationId: createIndication
      description: Creates an indication for the appointment
      tags:
        - indications
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/IndicationForm'
            encoding:
              file:
                contentType: image/png, image/jpeg, application/pdf
      responses:
        '201':
          description: Created indication
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '403':
          description: Only the patient or the doctor can create an indication for the appointment
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Appointment not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
  /appointments/{appointmentId}/indications/{indicationId}:
    get:
      operationId: getIndication
      description: Gets the indication
      tags:
        - indications
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: indicationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returned indication
          content:
            application/vnd.indication.v1+json:
              schema:
                allOf:
                  - $ref : '#/components/schemas/Indication'
        '403':
          description: Only the patient or the doctor can see the indication of the appointment
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Indication or appointment not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
...
components:
  schemas:
    CancelAppointmentForm:
      required:
        - description
        - status
      type: object
      properties:
        status:
          pattern: CANCELLED
          type: string
        description:
          maxLength: 1000
          minLength: 1
          type: string
          example: 'Tengo examen ese dia, mas adelante agendo otro turno'
...
    AppointmentForm:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2023-08-15"
        timeBlock:
          allOf:
            - $ref : '#/components/schemas/TimeBlock'
          example: "10:00"
        description:
          type: string
          minLength: 1
          maxLength: 100
          example: "Control rutinario"
        doctorId:
          type: integer
          format: int64
          example: 123
...
    IndicationForm:
      type: object
      required:
        - indications
      properties:
        indications:
          type: string
        file:
          type: string
          format: binary
...
    Error:
      type: object
      properties:
        message:
          type: string
    ValidationError:
      type: object
      properties:
        attribute:
          type: string
        invalidValue:
          type: string
        message:
          type: string
...
    Indication:
      type: object
      properties:
        appointment:
          type: string
          format: uri
          description: The URI for the related appointment.
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/appointments/12"
        creator:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20"
        date:
          type: string
          format: date
          example: "2023-06-26"
        description:
          type: string
          example: "Buen dia, le adjunto el resultado del estudio"
        file:
          type: string
          format: uri
          description: The URI for the file related to the indication.
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/appointments/12/files/1"
        id:
          type: integer
          format: int64
          example: 2
        self:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/appointments/12/indications/2"
...
    Appointment:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2023-05-20"
        description:
          type: string
          example: "Me duele el pie"
        cancelDescription:
          type: string
          example: null
        doctor:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20"
        id:
          type: integer
          format: int64
          example: 12
        patient:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/patients/21"
        self:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/appointments/12"
        status:
          type: string
          example: "COMPLETED"
        timeBlock:
          type: string
          example: "10:30"
...
    TimeBlock:
      type: string
      enum:
        - "00:00"
        - "00:30"
        - "01:00"
        - "01:30"
        - "02:00"
        - "02:30"
        - "03:00"
        - "03:30"
        - "04:00"
        - "04:30"
        - "05:00"
        - "05:30"
        - "06:00"
        - "06:30"
        - "07:00"
        - "07:30"
        - "08:00"
        - "08:30"
        - "09:00"
        - "09:30"
        - "10:00"
        - "10:30"
        - "11:00"
        - "11:30"
        - "12:00"
        - "12:30"
        - "13:00"
        - "13:30"
        - "14:00"
        - "14:30"
        - "15:00"
        - "15:30"
        - "16:00"
        - "16:30"
        - "17:00"
        - "17:30"
        - "18:00"
        - "18:30"
        - "19:00"
        - "19:30"
        - "20:00"
        - "20:30"
        - "21:00"
        - "21:30"
        - "22:00"
        - "22:30"
        - "23:00"
        - "23:30"
...